moved fixture to central config file
