+++ tests/components/geonetnz_volcano/test_init.py
@@ -1,33 +1,7 @@
 """Define tests for the GeoNet NZ Volcano general setup."""
 from asynctest import CoroutineMock, patch
-import pytest
 
 from homeassistant.components.geonetnz_volcano import DOMAIN, FEED
-from homeassistant.const import (
-    CONF_LATITUDE,
-    CONF_LONGITUDE,
-    CONF_RADIUS,
-    CONF_SCAN_INTERVAL,
-    CONF_UNIT_SYSTEM,
-)
-
-from tests.common import MockConfigEntry
-
-
-@pytest.fixture
-def config_entry():
-    """Create a mock GeoNet NZ Volcano config entry."""
-    return MockConfigEntry(
-        domain=DOMAIN,
-        data={
-            CONF_LATITUDE: -41.2,
-            CONF_LONGITUDE: 174.7,
-            CONF_RADIUS: 25,
-            CONF_UNIT_SYSTEM: "metric",
-            CONF_SCAN_INTERVAL: 300.0,
-        },
-        title="-41.2, 174.7",
-    )
 
 
 async def test_component_unload_config_entry(hass, config_entry):
